added new route /thankyou
