+++ src/app/upload/page.tsx
@@ -1,5 +1,6 @@
 "use client";
 
+import { useRouter } from "next/navigation";
 import { Suspense, useState } from "react";
 import { InboxOutlined } from "@ant-design/icons";
 import type { UploadProps } from "antd";
@@ -11,27 +12,9 @@ import { useLoading } from "@/hooks/useLoading";
 import { UploadResult, FileUploadAntD } from "@/types/upload";
 import { IMAGE_KIT_PUBLIC_KEY } from "@/constants";
 
-function ThankYouMessage() {
-  return (
-    <div className="flex flex-col justify-center items-center p-10 text-center">
-      <h3 className="text-green-600 text-2xl font-bold">
-        Thank you for your submission! Your project details and images have
-        successfully been submitted.
-      </h3>
-      <h1 className="font-semibold text-l m-4">
-        You’ll get a confirmation email with payment details soon.
-      </h1>
-      <h5 className="font-thin">
-        Payment is only required for the first project you submit. Any
-        additional submissions are free of charge.
-      </h5>
-    </div>
-  );
-}
-
 function ImageUploader() {
   const [files, setFiles] = useState<Array<FileUploadAntD>>([]);
-  const [submittedSuccess, setSubmittedSuccess] = useState<boolean>(false);
+  const router = useRouter();
   const [titleImageUid, setTitleImageUid] = useState<string | null>(null);
   const searchParams = useSearchParams();
   const email = searchParams.get("email");
@@ -44,7 +27,7 @@ function ImageUploader() {
     name: "file",
     multiple: true,
     style: { maxHeight: "10rem" },
-    disabled: submittedSuccess || loading,
+    disabled: loading,
     accept: ".jpg, .png, .avif, .gif, .webp",
     beforeUpload(file) {
       const isLt2M = file.size / 1024 / 1024 < 2;
@@ -81,7 +64,7 @@ function ImageUploader() {
           <Button
             size="small"
             className="ml-2"
-            disabled={loading || submittedSuccess}
+            disabled={loading}
             type={titleImageUid === file.uid ? "primary" : "default"}
             onClick={() => {
               setTitleImageUid(file.uid);
@@ -94,30 +77,9 @@ function ImageUploader() {
     },
   };
 
-  const resetForm = () => {
-    setFiles([]);
-    setTitleImageUid(null);
-    setSubmittedSuccess(false);
-  };
-
   return (
     <div className="min-h-screen flex flex-col justify-center items-center">
-      {submittedSuccess ? (
-        <>
-          <ThankYouMessage />
-          <Button
-            className="m-5 p-4 mt-5"
-            type="primary"
-            // haven't created a test for this yet or maybe we need can use the same test?
-            //data-testid="image-submit-again"
-            loading={loading}
-            onClick={resetForm}
-            >
-              Submit Another Project
-          </Button>
-        </>
-      ) : (
-        <>
+      <>
           <h2 className="font-bold text-3xl p-4">
             The Final Step! Upload all images associated with the project
           </h2>
@@ -185,7 +147,7 @@ function ImageUploader() {
                   })
                 );
                 console.log("everything uploaded", res);
-                setSubmittedSuccess(true);
+                router.push("/thankyou");
                 return res;
               } catch (err) {
                 console.error("failed upload", err);
@@ -214,7 +176,7 @@ function ImageUploader() {
                 type="primary"
                 data-testid="image-submit"
                 loading={loading}
-                disabled={loading || submittedSuccess}
+                disabled={loading}
                 htmlType="submit"
               >
                 Submit Images
@@ -222,7 +184,6 @@ function ImageUploader() {
             </div>
           </form>
         </>
-      )}
     </div>
   );
 }
